Move commit function down the component tree
